services/embeddings/requirements.txt
torch>=2.0.0
transformers>=4.45.0
qwen-vl-utils>=0.0.14
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
numpy>=1.24.0
